Move quote.h to public headers
