+++ src/node/rpc/node_frontend.h
@@ -7,6 +7,7 @@
 #include "ccf/http_query.h"
 #include "ccf/json_handler.h"
 #include "ccf/odata_error.h"
+#include "ccf/quote.h"
 #include "ccf/version.h"
 #include "consensus/aft/orc_requests.h"
 #include "crypto/certs.h"
@@ -15,7 +16,6 @@
 #include "enclave/reconfiguration_type.h"
 #include "frontend.h"
 #include "node/network_state.h"
-#include "node/quote.h"
 #include "node/rpc/jwt_management.h"
 #include "node/rpc/serialization.h"
 #include "node/session_metrics.h"
